Remove JDBCUtils. Move ColumnFetcherFactory into the Type Mappings. Alter SnowflakeJDBCUtils to be a derived JDBCValueSetter.
