+++ distribution/lib/Standard/Database/0.0.0-dev/src/Connection/Connection.enso
@@ -37,8 +37,6 @@ from project.Internal.Upload.Operations.Create import create_table_implementatio
 from project.SQL_Query import make_table_for_name
 
 polyglot java import java.lang.UnsupportedOperationException
-polyglot java import org.enso.database.JDBCUtils
-polyglot java import org.enso.database.fetchers.ColumnFetcherFactory
 
 type Connection
     ## ---
@@ -103,7 +101,7 @@ type Connection
     databases : Vector Text
     databases self =
         self.jdbc_connection.with_metadata metadata->
-            Vector.from_polyglot_array <| JDBCUtils.readTextColumn metadata.getCatalogs "TABLE_CAT"
+            Vector.from_polyglot_array <| self.type_mapping.column_fetcher_factory.readTextColumn metadata.getCatalogs "TABLE_CAT"
 
     ## ---
        icon: metadata
@@ -133,7 +131,7 @@ type Connection
     schemas : Vector Text
     schemas self =
         self.jdbc_connection.with_metadata metadata->
-            Vector.from_polyglot_array <| JDBCUtils.readTextColumn metadata.getSchemas "TABLE_SCHEM"
+            Vector.from_polyglot_array <| self.type_mapping.column_fetcher_factory.readTextColumn metadata.getSchemas "TABLE_SCHEM"
 
     ## ---
        icon: metadata
@@ -163,7 +161,7 @@ type Connection
     table_types : Vector Text
     table_types self =
         self.jdbc_connection.with_metadata metadata->
-            Vector.from_polyglot_array <| JDBCUtils.readTextColumn metadata.getTableTypes "TABLE_TYPE"
+            Vector.from_polyglot_array <| self.type_mapping.column_fetcher_factory.readTextColumn metadata.getTableTypes "TABLE_TYPE"
 
     ## ---
        group: Standard.Base.Metadata
@@ -207,7 +205,7 @@ type Connection
             _ : Vector -> types
             _ -> [types]
         tm = self.type_mapping
-        result = get_tables_advanced self.jdbc_connection name_like database schema types_vector tm.sql_type_to_value_type tm.column_fetcher_factory all_fields
+        result = get_tables_advanced self.jdbc_connection name_like database schema types_vector tm.column_fetcher_factory tm.sql_type_to_value_type tm.column_fetcher_factory all_fields
         case include_hidden of
             True -> result
             False ->
@@ -357,7 +355,7 @@ type Connection
             # And finally, materialize the results.
             tm = self.type_mapping
             SQL_Warning_Helper.process_warnings stmt <|
-                result_set_to_table rs tm.sql_type_to_value_type tm.column_fetcher_factory type_overrides last_row_only
+                result_set_to_table rs tm.column_fetcher_factory tm.sql_type_to_value_type type_overrides last_row_only
 
     ## ---
        private: true
@@ -377,7 +375,7 @@ type Connection
     read_text_column : Text -> Text -> Vector Text
     read_text_column self query column_name = self.jdbc_connection.with_connection connection->
         result_set = connection.createStatement.executeQuery query
-        Vector.from_polyglot_array <| JDBCUtils.readTextColumn result_set column_name
+        Vector.from_polyglot_array <| self.type_mapping.column_fetcher_factory.readTextColumn result_set column_name
 
     ## ---
        private: true
@@ -404,7 +402,7 @@ type Connection
                 _check_statement_is_allowed self stmt
                 rs = stmt.executeQuery
                 tm = self.type_mapping
-                table = result_set_to_table rs tm.sql_type_to_value_type tm.column_fetcher_factory row_limit=(limit.rows_to_read.if_nothing -1)
+                table = result_set_to_table rs tm.column_fetcher_factory tm.sql_type_to_value_type row_limit=(limit.rows_to_read.if_nothing -1)
                 limit.attach_warning table
 
     ## ---
@@ -576,7 +574,7 @@ type Connection
             used_schema = if schema_name == "" then Nothing else schema_name
             rs = metadata.getPrimaryKeys Nothing used_schema table_name
             tm = self.type_mapping
-            keys_table = result_set_to_table rs tm.sql_type_to_value_type tm.column_fetcher_factory
+            keys_table = result_set_to_table rs tm.column_fetcher_factory tm.sql_type_to_value_type
             # The names of the columns are sometimes lowercase and sometimes uppercase, so we do a case insensitive select first.
             selected = keys_table.select_columns ["COLUMN_NAME", "KEY_SEQ"] case_sensitivity=Case_Sensitivity.Insensitive reorder=True
             key_column_names = selected.sort 1 . at 0 . to_vector
@@ -659,11 +657,11 @@ private _check_statement_is_allowed connection stmt =
    private: true
    ---
    Get a metadata table for tables.
-get_tables_advanced jdbc_connection name_like:Text|Nothing database:Text|Nothing schema:Text|Nothing types:Nothing|Vector sql_type_to_value_type=(_->Nothing) column_fetcher_factory=ColumnFetcherFactory.DEFAULT all_fields:Boolean=False =
+get_tables_advanced jdbc_connection name_like:Text|Nothing database:Text|Nothing schema:Text|Nothing types:Nothing|Vector  column_fetcher_factory sql_type_to_value_type=(_->Nothing) all_fields:Boolean=False =
     name_dict = Dictionary.from_vector [["TABLE_CAT", "Database"], ["TABLE_SCHEM", "Schema"], ["TABLE_NAME", "Name"], ["TABLE_TYPE", "Type"], ["REMARKS", "Description"], ["TYPE_CAT", "Type Database"], ["TYPE_SCHEM", "Type Schema"], ["TYPE_NAME", "Type Name"]]
     jdbc_connection.with_metadata metadata->
         table = Managed_Resource.bracket (metadata.getTables database schema name_like types) .close result_set->
-            result_set_to_table result_set sql_type_to_value_type column_fetcher_factory
+            result_set_to_table result_set column_fetcher_factory sql_type_to_value_type
         renamed = table.rename_columns name_dict error_on_missing_columns=False on_problems=..Ignore
         if all_fields then renamed else
             renamed.select_columns ["Database", "Schema", "Name", "Type", "Description"]
