+++ distribution/lib/Standard/Database/0.0.0-dev/src/Internal/Helpers.enso
@@ -9,8 +9,6 @@ import project.DB_Table.DB_Table
 import project.Internal.IR.Internal_Column.Internal_Column
 from project.Errors import Unsupported_Database_Operation
 
-polyglot java import java.util.regex.Pattern
-
 ## ---
    private: true
    ---
