+++ distribution/lib/Standard/Database/0.0.0-dev/src/Internal/Postgres/Helpers.enso
@@ -5,6 +5,8 @@ import Standard.Base.Errors.Illegal_State.Illegal_State
 import project.Internal.JDBC_Connection.JDBC_Connection
 from project.Errors import Unsupported_Database_Encoding
 
+polyglot java import org.enso.database.fetchers.ColumnFetcherFactory
+
 ## ---
    private: true
    ---
@@ -19,7 +21,7 @@ get_encoding_name jdbc_connection =
    ---
 get_pragma_value : JDBC_Connection -> Text -> Any
 get_pragma_value jdbc_connection sql =
-    table = jdbc_connection.raw_read_statement sql
+    table = jdbc_connection.raw_read_statement sql ColumnFetcherFactory.DEFAULT
     if table.row_count != 1 then
         Panic.throw (Illegal_State.Error "Unexpected amount of results to internal query: "+sql+"; expected 1 but got "+table.row_count.to_text+" rows.")
     if table.columns.length != 1 then
