+++ distribution/lib/Standard/Database/0.0.0-dev/src/Internal/Postgres/Postgres_Type_Mapping.enso
@@ -14,6 +14,7 @@ import project.SQL.SQL_Type
 from project.Errors import Unsupported_Database_Type
 
 polyglot java import java.sql.Types
+polyglot java import org.enso.database.fetchers.ColumnFetcherFactory
 polyglot java import org.enso.database.JDBCValueSetter
 
 ## ---
@@ -100,9 +101,10 @@ type Postgres_Type_Mapping
        private: true
        advanced: true
        ---
-       Gets the ColumnFetcherFactory to be used for this backend.
+       Gets the ColumnFetcherFactory to be used for Postgres.
     column_fetcher_factory : Any
-    column_fetcher_factory = SQL_Type_Mapping.default_column_fetcher_factory
+    column_fetcher_factory =
+        ColumnFetcherFactory.DEFAULT
 
     ## ---
        private: true
