+++ distribution/lib/Standard/Microsoft/0.0.0-dev/src/Internal/SQLServer_Type_Mapping.enso
@@ -19,7 +19,6 @@ import project.Internal.SQLServer_Error_Mapper.SQLServer_Error_Mapper
 
 polyglot java import java.sql.Types
 polyglot java import org.enso.database.fetchers.ColumnFetcherFactory
-polyglot java import org.enso.database.JDBCUtils
 polyglot java import org.enso.database.JDBCValueSetter
 
 ## ---
@@ -123,7 +122,7 @@ type SQLServer_Type_Mapping
        private: true
        advanced: true
        ---
-       Gets the ColumnFetcherFactory to be used for this backend.
+       Gets the ColumnFetcherFactory to be used for SQL Server.
     column_fetcher_factory : Any
     column_fetcher_factory =
         ColumnFetcherFactory.DEFAULT
