+++ std-bits/database/src/main/java/org/enso/database/JDBCValueSetter.java
@@ -23,7 +23,7 @@ public class JDBCValueSetter {
     return new JDBCValueSetter(databaseName);
   }
 
-  private JDBCValueSetter(String databaseName) {
+  protected JDBCValueSetter(String databaseName) {
     this.databaseName = databaseName;
   }
